Test for 'Email for access options' link
